% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avonet_extract.R
\name{avonet_extract_interspecies}
\alias{avonet_extract_interspecies}
\title{Extract data from interspecies mean tibbles (i.e. $birdlife, $ebird, $birdtree); bypasses avonet_load()}
\usage{
avonet_extract_interspecies(filename, taxon_system = "birdlife", traits = NULL)
}
\arguments{
\item{filename}{A character describing the file path to the downloaded avonet dataset (same as filename in avonet_install()).}

\item{taxon_system}{A character defining the taxon classification system to use. Define as 'birdlife', 'ebird', or 'birdtree'; accesses the specified tibble.}

\item{traits}{A character vector describing the non-taxonomy variables to extract (usually traits).}
}
\value{
A tibble with the extracted data from $birdlife, $ebird, or $birdtree
}
\description{
Extract data from interspecies mean tibbles (i.e. $birdlife, $ebird, $birdtree); bypasses avonet_load()
}
\examples{
avonet_install(filename = "raw_data/avonet_dataset.xlsx")
avonet_extract_interspecies(filename = "raw_data/avonet_dataset.xlsx", traits = c("tail.length", "wing.length"))
}
