% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avonet_install.R
\name{avonet_install}
\alias{avonet_install}
\title{Installs avonet dataset}
\usage{
avonet_install(filename)
}
\arguments{
\item{filename}{A character defining the file destination}
}
\value{
Downloads avonet dataset at specified location
}
\description{
Installs avonet dataset
}
\examples{
avonet_install(filename = "raw_data/avonet_dataset.xlsx")
}
