% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avonet_summary.R
\name{avonet_summary}
\alias{avonet_summary}
\title{Apply summary statistics to extracted data}
\usage{
avonet_summary(data, group = c("species"), stats = c("mean"), tidy = FALSE)
}
\arguments{
\item{data}{A tibble produced by an avonet_extract_*() function}

\item{group}{A character vector to summarise by (e.g. c("species", "sex"))}

\item{stats}{A character vector of summary statistics (e.g. c("mean", "median", "sd"))}

\item{tidy}{A logical value defining tibble format; FALSE returns long, visually preferable form while TRUE returns wide, statistically preferable form}
}
\value{
A tibble summarising extracted data
}
\description{
Apply summary statistics to extracted data
}
\examples{
avonet_install(filename = "raw_data/avonet_dataset.xlsx")
extract <- avonet_extract_intraspecies(filename = "raw_data/avonet_dataset.xlsx", taxon_system = "birdlife", traits = c("sex", ".length"), clean = "all")
avonet_summary(data = extract, group = c("species", "sex"), stats = c("mean", "median", "max", "min", "sd"), tidy = FALSE)
}
